Apply @mofeing suggestions from code review

+++ src/Counters.jl
@@ -6,4 +6,4 @@ flops(a, b, size, keep) = prod(ind -> size[ind], flatten((a ∪ b, ∩(keep, a, 
 
 rank(a, b, size, keep) = length(symdiff(a, b) ∪ ∩(keep, a, b))
 
-removedsize(a, b, size, keep) = prod(size[ind] for ind ∈ symdiff(a, b) ∪ (keep ∩ a) ∪ (keep ∩ b); init=Float64(1)) - prod(size[ind] for ind ∈ a) - prod(size[ind] for ind ∈ b)
+removedsize(a, b, size, keep) = prod(size[ind] for ind ∈ symdiff(a, b) ∪ (keep ∩ a) ∪ (keep ∩ b); init=1) - prod(size[ind] for ind ∈ a) - prod(size[ind] for ind ∈ b)
